:recycle: [Refactor]: 테스트용 인증/인가 관련 코드 분리
