+++ src/main/java/com/avab/avab/service/AuthService.java
@@ -8,8 +8,6 @@ public interface AuthService {
 
     OAuthResponse kakaoLogin(String code);
 
-    OAuthResponse kakaoLoginLocal(String code);
-
     TokenRefreshResponse refresh(String refreshToken);
 
     void logout(Long userId);
